Implement a query_ skeleton with direct-sqlite
This commit puts together enough plumbing to run a query using
direct-sqlite prepare/step/columns and stick it into a Result that can
be used for converting values to Haskell.  This is just the bare
minimum to get some data through to the user.

There's plenty of improvement in functionality, cleanliness and
performance.

The goal here is to have some data going through every module, so that
I can get basic tests in place.