Add test file to check that imports are ok (#10, #9)
Make sure some basic types are exported from Database.SQLite.Simple
so that users of this library don't necessarily need to import
too many modules just to get going.

Adds a source file that's compiled as part of tests that sort of
tests that the most of the necessary things come from a single
module.

Bump up version number in preparation for the next release.