Use doFold instead of stepStmt to build up the result list
This should produce less intermediate allocations as a single row at a
time is read from the database and converted to the resulting type.
In the previous implementation, all the result rows were stored into a
single list of SQLData fields which was then converted to another list
of the target type.