Dataype generic implementation of {From,To}Row
This is my first time ever doing data-type generic programming, so
please review this carefully.  I conferred with the documentation in
base[1] and I also looked at the analoguous definition in the package
`binary`[2].  It was surprisingly easy to implement, but perhaps it
should be tested some more before merging in.  I've just quickly
tested it with the code-base I'm working on.

Note that it the current implementation can only derive the generic
instance for product-types (not sum-types).

[1]: http://hackage.haskell.org/package/base-4.12.0.0/docs/GHC-Generics.html#g:10
[2]: http://hackage.haskell.org/package/binary